# public/openapi.yaml
openapi: 3.1.0
info:
  title: BasaltSurge API
  version: 1.0.0
  contact:
    name: BasaltSurge Support
    url: https://surge.basalthq.com
    email: support@basalthq.com
  description: >
    Public API surface for BasaltSurge under a hardened posture: Azure Front
    Door Premium (AFD) gateway → Azure API Management (APIM).

    All developer-facing endpoints require an APIM subscription key header
    `Ocp-Apim-Subscription-Key` and use the APIM custom domain. AFD is optional
    as a fallback.

    APIM policy permits requests with a valid subscription key and may also
    accept AFD-originated requests carrying `x-edge-secret` if AFD is
    configured. Direct backend origins are not supported.

    Wallet identity is resolved at the gateway based on subscription;
    client-supplied `x-wallet` header and `wallet` parameters are ignored.

    Health endpoint: `GET /healthz` returns 200 JSON for probes; does not
    require a subscription.

    Note: Some endpoints are read-only or environment-gated in production.
    Admin/sensitive routes (JWT-only) are not part of this spec.
  x-guidance: >
    Welcome to BasaltSurge's Agentic Commerce API.

    To use this API:

    1. Non-paid resources are fully accessible with standard API keys or without
    auth depending on the path.

    2. Payment-gated resources under /api/x402/ must be paid for. For example,
    POST /api/x402/orders to discover the pricing required. The server will
...

...

...
    the x402 challenge there.
...
servers:
...
security:
  - apimKey: []
...
paths:
...
  /api/x402/subscribe:
    post:
      summary: Subscribe to API Tier via x402 Payment
      operationId: x402Subscribe
      tags:
        - Subscriptions
      description: >
        x402-gated developer subscription endpoint. Returns 402 with payment
        challenge for paid tiers (pro=$399, enterprise=$500). Starter tier is
        free and does not require payment. Once paid, returns an API key.
      security:
        - x402: []
      x-payment-info:
        price:
          mode: dynamic
          currency: USD
          min: '0.00'
          max: '5000.00'
        protocols:
          - x402: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - planKey
              properties:
                planKey:
                  type: string
                  enum:
                    - starter
                    - pro
                    - enterprise
                  description: 'Pricing: starter=$0, pro=$399, enterprise=$500'
      responses:
        '200':
          description: Subscription activated (API key issued)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  plan:
                    type: string
                  priceUsd:
                    type: number
                  subscriptionId:
                    type: string
                  apiKey:
                    type: string
                  message:
                    type: string
        '402':
          description: Payment Required – x402 challenge issued
          headers:
            Payment-Required:
              description: Base64 encoded payload
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  x402Version:
                    type: integer
                  error:
                    type: string
                  accepts:
                    type: array
                    items:
                      type: object
...
components:
  securitySchemes:
    apimKey:
      type: apiKey
      in: header
      name: Ocp-Apim-Subscription-Key
      description: >
        Azure API Management subscription key. Required for developer-facing
        endpoints.

        Obtain your key via BasaltSurge Admin → API Subscriptions (or directly
        in APIM).
    x402:
      type: apiKey
      in: header
      name: x-payment
      description: x402 Agentic Payment Standard Proof (Base64 payload)
...